requirements.txt
# Modular Streamlit MCP Application Requirements

# Core dependencies
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
